Update lab 2.cpp
I commented out part of the code you used for testing your functions, so I could test the implementation with the switch menu. But I didn't delete or alter any of your original stuff. It's all there. :)

stringNCopy is the only piece of code that still needs work.

I included: 
-Program description + Pseudocode (top of file);
-Function prototypes;
-C-string variables + const int for length;
-Implemented switch menu with available functions;
-Added functions for single and double string input;
-Added missing documentation to the my code;

To be discussed:
-Do we need the single string input functions I added?
-Do we need to improve the messages being printed out to the user?
-What else is missing?