Update lab 2.cpp
-Improved my stringAdd function.

I run a few tests and had trouble with the following:
- Menu accepting char instead of int (see comment next to variable)
- Both lines with cin.ignore(10000, '\n') caused trouble (I didn't remove them here, but left a comment. Had to remove when tested)
- While loops to print c-string inside case 2 and 3 were not necessary. I could print it directly with the variable. You must have had a reason to use it. (I didn't remove them, but left a comment)
- Marked 2 variables that aren't used on the main scope to be DELETED.

Other than that, everything is working well  and the functions are perfect . I'm concerned about the "One of the menu options should be to continue the test or exit the program" posted on the lab description, since we have a do-while loop instead. 

I'd say we have 95% (or more) of our lab finished. Thanks for the awesome job! :)